+++ stories/primer/alpha/truncate_stories.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "primer/alpha/button_marketing"
+
+class Primer::Alpha::TruncateStories < ViewComponent::Storybook::Stories
+  layout "storybook_preview"
+
+  story(:truncate) do
+    content do |component|
+      component.text() { "Truncate text" }
+    end
+  end
+end